\bra and \ket
i did part of the replacing in vim (before I realized I wasn't even
using macros etc. and switched back to find and replace in texstudio)
and in the process vim fixed some whitespace based on my preferences -
annoying when the rest of the document was written with a different
editor!

additionally fixed a couple (x)_i -> {(x)}_i things and other lints
since I had a linter (wonder if texstudio has lints)