finite group dumps from single cgate
i should change the format so that it prints the word, then the order,
then the matrix, which would require moving the order calculation to the
print function rather than the print test. not that interesting to me
right now, that is polish, for something that will barely be an
appendix, but merely a hyperlink.